Move the types module up a level and consolidate with enum. Introduce slight hack for rules hook.
