+++ src/sqlfluff/core/config/toml.py
@@ -8,12 +8,12 @@ if sys.version_info >= (3, 11):
 else:  # pragma: no cover
     import toml as tomllib
 
-from sqlfluff.core.config.types import ConfigMappingType
 from sqlfluff.core.helpers.dict import (
     NestedDictRecord,
     iter_records_from_nested_dict,
     records_to_nested_dict,
 )
+from sqlfluff.core.types import ConfigMappingType
 
 T = TypeVar("T")
