+++ src/sqlfluff/core/config/validate.py
@@ -1,8 +1,8 @@
 """Methods for validating config dicts."""
 
 from sqlfluff.core.config.removed import validate_config_dict_for_removed
-from sqlfluff.core.config.types import ConfigMappingType
 from sqlfluff.core.errors import SQLFluffUserError
+from sqlfluff.core.types import ConfigMappingType
 
 ALLOWABLE_LAYOUT_CONFIG_KEYS = (
     "spacing_before",
