+++ src/sqlfluff/core/rules/base.py
@@ -48,6 +48,7 @@ from sqlfluff.core.rules.context import RuleContext
 from sqlfluff.core.rules.crawlers import BaseCrawler
 from sqlfluff.core.rules.fix import LintFix
 from sqlfluff.core.templaters.base import TemplatedFile
+from sqlfluff.core.types import RuleType
 
 # Best solution for generic types on older python versions
 # https://github.com/python/typeshed/issues/7855
@@ -330,8 +331,8 @@ class RuleMetaclass(type):
                 break
 
         class_dict = RuleMetaclass._populate_docstring(name, class_dict)
-        # Don't try and infer code and description for the base class
-        if bases:
+        # Don't try and infer code and description for the base classes
+        if name not in ("BaseRule", "RuleType"):
             class_dict = RuleMetaclass._populate_code_and_description(name, class_dict)
         # Validate rule names
         rule_name = class_dict.get("name", "")
@@ -350,7 +351,7 @@ class RuleMetaclass(type):
         return super().__new__(mcs, name, bases, class_dict)  # type: ignore
 
 
-class BaseRule(metaclass=RuleMetaclass):
+class BaseRule(RuleType, metaclass=RuleMetaclass):
     """The base class for a rule.
 
     Args:
